Rename fragments to match codelab
