+++ app/src/main/java/com/example/wordsapp/WordListFragment.kt
@@ -16,35 +16,49 @@
 package com.example.wordsapp
 
 import android.os.Bundle
-import android.view.*
-import androidx.core.content.ContextCompat
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.wordsapp.databinding.FragmentWordListBinding
 
 /**
- * Entry fragment for the app. Displays a [RecyclerView] of letters.
+ * Displays a [RecyclerView] of words with search buttons to look them up.
  */
 class WordListFragment : Fragment() {
+    private lateinit var letterId: String
+
     private var _binding: FragmentWordListBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
     private val binding get() = _binding!!
 
-    private lateinit var recyclerView: RecyclerView
-    // Keeps track of which LayoutManager is in use for the [RecyclerView]
-    private var isLinearLayoutManager = true
+
+    /**
+     * Provides global access to these variables from anywhere in the app
+     * via DetailListFragment.<variable> without needing to create
+     * a DetailListFragment instance.
+     */
+    companion object {
+        val LETTER = "letter"
+        val SEARCH_PREFIX = "https://www.google.com/search?q="
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setHasOptionsMenu(true)
+
+        // Retrieve the LETTER from the Fragment arguments
+        arguments?.let {
+            letterId = it.getString(LETTER).toString()
+        }
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Retrieve and inflate the layout for this fragment
@@ -54,80 +68,14 @@ class WordListFragment : Fragment() {
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        recyclerView = binding.recyclerView
-        // Sets the LayoutManager of the recyclerview
-        // On the first run of the app, it will be LinearLayoutManager
-        chooseLayout()
-    }
+        val recyclerView = binding.recyclerView
+        recyclerView.layoutManager = LinearLayoutManager(requireContext())
+        recyclerView.adapter = WordAdapter(letterId, requireContext())
 
-    /**
-     * Sets the LayoutManager for the [RecyclerView] based on the desired orientation of the list.
-     *
-     * Notice that because the enclosing class has changed from an Activity to a Fragment,
-     * the signature of the LayoutManagers has to slightly change.
-     */
-    private fun chooseLayout() {
-        if (isLinearLayoutManager) {
-            recyclerView.layoutManager = LinearLayoutManager(context)
-        } else {
-            recyclerView.layoutManager = GridLayoutManager(context, 4)
-        }
-        recyclerView.adapter = LetterAdapter()
-    }
-
-    private fun setIcon(menuItem: MenuItem?) {
-        if (menuItem == null)
-            return
-
-        // Set the drawable for the menu icon based on which LayoutManager is currently in use
-
-        // An if-clause can be used on the right side of an assignment if all paths return a value.
-        // The following code is equivalent to
-        // val context = this.requireContext()
-        // if (isLinearLayoutManager)
-        //     menu.icon = ContextCompat.getDrawable(context, R.drawable.ic_grid_layout)
-        // else menu.icon = ContextCompat.getDrawable(context, R.drawable.ic_linear_layout)
-        menuItem.icon =
-            if (isLinearLayoutManager)
-                ContextCompat.getDrawable(this.requireContext(), R.drawable.ic_grid_layout)
-            else ContextCompat.getDrawable(this.requireContext(), R.drawable.ic_linear_layout)
-    }
-
-    /**
-     * Initializes the [Menu] to be used with the current [Activity]
-     *
-     * The signature of this function has also changed slightly when changing from
-     * an Activity to a Fragment
-     */
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.layout_menu, menu)
-
-        val layoutButton = menu.findItem(R.id.action_switch_layout)
-        // Calls code to set the icon based on the LinearLayoutManager of the RecyclerView
-        setIcon(layoutButton)
-    }
-
-    /**
-     * Determines how to handle interactions with the selected [MenuItem]
-     */
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
-            R.id.action_switch_layout -> {
-                // Sets isLinearLayoutManager (a Boolean) to the opposite value
-                isLinearLayoutManager = !isLinearLayoutManager
-                // Sets layout and icon
-                chooseLayout()
-                setIcon(item)
-
-                return true
-            }
-            //  Otherwise, do nothing and use the core event handling
-
-            // when clauses require that all possible paths be accounted for explicitly,
-            //  for instance both the true and false cases if the value is a Boolean,
-            //  or an else to catch all unhandled cases.
-            else -> super.onOptionsItemSelected(item)
-        }
+        // Adds a [DividerItemDecoration] between items
+        recyclerView.addItemDecoration(
+            DividerItemDecoration(context, DividerItemDecoration.VERTICAL)
+        )
     }
 
     /**
