+++ app/src/main/res/navigation/nav_graph.xml
@@ -17,23 +17,21 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/wordList">
-
+    app:startDestination="@id/letterListFragment">
     <fragment
-        android:id="@+id/wordList"
+        android:id="@+id/letterListFragment"
+        android:name="com.example.wordsapp.LetterListFragment"
         android:label="@string/app_name"
-        android:name="com.example.wordsapp.WordListFragment"
-        tools:layout="@layout/fragment_word_list">
+        tools:layout="@layout/fragment_letter_list">
         <action
-            android:id="@+id/action_wordList_to_detailList"
-            app:destination="@id/detailList"
-            app:popUpToInclusive="true" />
+            android:id="@+id/action_letterListFragment_to_wordListFragment"
+            app:destination="@id/wordListFragment" />
     </fragment>
     <fragment
-        android:id="@+id/detailList"
-        android:name="com.example.wordsapp.DetailListFragment"
+        android:id="@+id/wordListFragment"
+        android:name="com.example.wordsapp.WordListFragment"
         android:label="@string/fragment_label"
-        tools:layout="@layout/fragment_detail_list">
+        tools:layout="@layout/fragment_word_list">
         <argument
             android:name="letter"
             app:argType="string" />
